requirement.txt
aioredis==1.3.1
amqp==2.6.1
anyascii==0.1.7
appdirs==1.4.3
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
bleach==3.3.0
bootstrap3-datetime==2.4
CacheControl==0.12.6
cachetools==4.2.1
captcha==0.3
celery==4.4.7
certifi==2019.11.28
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
conditional==1.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.2.1
cycler==0.10.0
daphne==2.3.0
decorator==4.4.2
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.1
django-annoying==0.10.4
django-appconf==1.0.4
django-captcha==1.0.0
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-classy-tags==2.0.0
django-cms==3.8.0
django-cms-forms==0.1.2
django-compressor==2.4
django-cors-headers==3.5.0
django-countries==7.0
django-crispy-forms==1.10.0
django-email-extras==0.3.4
django-filer==2.0.2
django-filter==2.4.0
django-form-builder==0.16.3
django-formfactory==0.2.3
django-formtools==2.2
django-hitcount==1.3.0
django-ipware==3.0.2
django-js-asset==1.2.2
django-json-field==0.5.7
django-markdown-app==0.9.3.1
django-modelcluster==5.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-npm==1.0.0
django-polymorphic==3.0.0
django-recaptcha==2.0.6
django-rest-framework==0.1.0
django-sekizai==2.0.0
django-select2==7.4.2
django-settings-export==1.2.1
django-simplemde==0.1.3
django-taggit==1.3.0
django-taggit-labels==0.7.0
django-timezone-field==4.0
django-treebeard==4.4
django-upload-validator==1.1.5
django-user-agents==0.4.0
django-yamlfield==1.0.3
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.1
djangosaml2==0.50.0
dlib==19.22.0
docutils==0.16
draftjs-exporter==2.1.7
dulwich==0.20.14
easy-thumbnails==2.7.1
entrypoints==0.3
et-xmlfile==1.0.1
ffmpeg==1.4
filesig==0.3.0.post1
future==0.18.2
gevent==21.1.2
google-auth==1.25.0
greenlet==1.0.0
h11==0.12.0
h2==3.2.0
hashids==1.3.1
hiredis==1.1.0
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.3
incremental==17.5.0
ipaddr==2.2.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kombu==4.6.11
kubernetes==12.0.1
l18n==2020.6.1
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.4.1
MediaInfo==0.0.9
mistune==0.8.4
moviepy==1.0.3
msal==1.7.0
msgpack==1.0.2
mysqlclient==2.0.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
notifications-plugin==0.1.2
npm==0.1.1
numpy==1.19.4
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.4.0.46
openpyxl==3.0.6
optional-django==0.1.0
packaging==20.3
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
priority==1.3.0
proglog==0.1.9
progress==1.5
prometheus-client==0.10.0
prompt-toolkit==3.0.8
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pymediainfo==5.0.3
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.17.3
pysaml2==6.3.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-gnupg==0.4.6
python-magic==0.4.18
python3-openid==3.2.0
pytoml==0.1.21
pytools==2020.4.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
rcssmin==1.0.6
redis==3.5.3
reportlab==3.5.59
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rjsmin==1.1.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sleekxmpp==1.3.3
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
sphinx-me==0.3
sqlparse==0.3.1
tablib==3.0.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
Twisted==21.2.0
txaio==20.12.1
tzlocal==2.1
ua-parser==0.10.0
Unidecode==1.1.2
urllib3==1.25.8
user-agents==2.2.0
uvicorn==0.13.4
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
webkit2png==0.8.2
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
widgetsnbextension==3.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
yarn==0.0.9
zope.event==4.5.0
zope.interface==5.2.0
